requirements.txt
groq
google-generativeai>=0.7.0
tavily-python>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.0.0
streamlit>=1.35.0
supabase>=2.0.0
